requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
colorama==0.4.6
constantly==23.10.4
coverage==7.9.1
cryptography==45.0.4
daphne==4.2.0
Django==5.2.3
djangorestframework==3.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
pyOpenSSL==25.1.0
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-decouple==3.8
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.0
tzdata==2025.2
zope.interface==7.2
